// src/interfaces/LoanObject.ts
interface LoanState {
  amountLoan: number;
  typeLoan: string;
  dateConcession: string;
  numberOfPaymentsLoan: number;
  quotaLoan: number;
  stateLoan: string;
  interestLoan: number;
  administrationAmountLoan: number;
  totalAmountLoan: number;
  secureAmountLoan: number;
}

export default LoanState;